feat: add coming soon default page for farming page
