+++ app/routes/__index/farming-coming-soon/$slug.tsx
