+++ app/routes/__index/farming-coming-soon/detail.tsx
